Establish canonical empty Cons
The API equates (*Cons)(nil) to &Cons{} (the nil and empty Cons)
but generates the latter where applicable.

This decision was made because the empty Cons is simpler to work with
and supports modification via Visitor callback (a future goal).

The Printer will output () for both nil and empty Cons.
All established components should test for equality of the Cons in
their API.